+++ inst/examples/ess.R
@@ -23,11 +23,20 @@ round(sum(ab_matched))
 
 ess(bmix, method="morita")
 
+# One may also calculate the ESS on the logit scale, which
+# gives slightly different results due to the parameter
+# transformation, e.g.:
+prior_logit <- mixnorm(c(1, log(5/15), sqrt(1/5 + 1/15)))
+ess(prior_logit, family=binomial)
+
+bmix_logit <- mixnorm(rob=c(0.2, 0, 2), inf=c(0.8, log(10/2), sqrt(1/10 + 1/2)))
+ess(bmix_logit, family=binomial)
+
 # Predictive consistency of elir
 n_forward <- 1E1
 bmixPred <- preddist(bmix, n=n_forward)
 pred_samp <- rmix(bmixPred, 1E2)
-# use more samples here for greater accuracy
+# use more samples here for greater accuracy, e.g.
 # pred_samp <- rmix(bmixPred, 1E3) 
 pred_ess <- sapply(pred_samp, function(r) ess(postmix(bmix, r=r, n=n_forward), "elir") )
 ess(bmix, "elir")
@@ -40,9 +49,14 @@ ess(nmix, "elir")
 
 ess(nmix, "moment")
 
-## the reference scale determines the ESS
+# the reference scale determines the ESS
 sigma(nmix) <- 20
 ess(nmix)
+
+# we may also interpret normal mixtures as densities assigned to
+# parameters of a logit transformed response rate of a binomial
+nmix_logit <- mixnorm(c(1, logit(1/4), 2/sqrt(10)))
+ess(nmix_logit, family=binomial)
 
 # Gamma mixture example
 gmix <- mixgamma(rob=c(0.3, 20, 4), inf=c(0.7, 50, 10))