+++ NEWS.md
@@ -1,3 +1,18 @@
+# RBesT 1.8-0 - January 3rd, 2025
+
+## Enhancements
+
+* Enable ESS calculation for normal mixture densities when used in the
+  context of a standard one-parameter exponential family through the
+  new `family` argument.
+
+## Bugfixes
+
+* Correct boundary behavior of `BinaryExactCI` function whenever no
+  responses or no non-responses are observed. Fixes issue #21.
+* Stabilize internal beta mixture information function, which corrects
+  unstable ESS ELIR computations. Addresses issue #22.
+
 # RBesT 1.7-4 - November 21st, 2024
 
 ## Enhancements